Allow changing name of template being previewed
Allow the template being replaced to be one with a different name than the
one being edited. This allows, for example, Template:X/sandbox to be
edited, but the change previewed as if it were Template:X. For simplicity,
it is hidden by default, but custom JavaScript or similar techniques allow
it to be used.

Change-Id: Ib7a6b65d0caa58e9a39e8e6dbb647a911a32c780