Fix location of SpecialTemplateSandbox.php
I1736e76a8 was merged without noticing the incorrect directory this file
was being moved to.

With the current lack of PHP namespaces in this extension, it would be
closer to PSR-4 to name the file "includes/SpecialTemplateSandbox.php".

If we do at some point add namespacing and were to make the class be
something like "TemplateSandbox\Specials\SpecialTemplateSandbox", the
correct filename would then be "includes/Specials/SpecialTemplateSandbox.php"
(note the difference in case).

Change-Id: I1f7ff4be5c417e1378683227233ed60a010da66c