Show an error message if the template was not used during parse
When previewing a page with TemplateSandbox, if the modified template
was not used when parsing the specified page, show an error message.
This is implemented for VE and the core wikitext editor.

The UI is identical to that used if the specified page does not exist.

Also, in the core wikitext editor, wrap all errors with an error box,
like how errors are shown in EditPage. Previously, the error message
was unstyled.

The feature is not yet implemented in live preview mode.

Bug: T279736
Change-Id: I6065f6092fab4e0a2baf13d8304c8533c7c90196