Fix accidentally removed stylesheet
Followup I682f5747c5fce437568200f0ec95e98ea08ae96e

I wanted to combine the combine the stylesheet with the script module,
but that causes it to load later, causing a layout shift in the HTML
generated by the server. So this is basically a revert, except that I've
renamed ext.TemplateSandbox.top to ext.TemplateSandbox.styles.

Bug: T279736
Change-Id: Iea8d849c392b60f93aa4d4cc98389a2cffd92ee5